+++ src/frontend/apps/desk/src/features/teams/member-management/components/MemberGrid.tsx
@@ -1,19 +1,24 @@
+import { zodResolver } from '@hookform/resolvers/zod';
 import {
   Button,
   DataGrid,
+  Input,
   SortModel,
   usePagination,
 } from '@openfun/cunningham-react';
 import React, { useEffect, useState } from 'react';
+import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
+import { z } from 'zod';
 
 import IconUser from '@/assets/icons/icon-user.svg';
-import { Box, Card, TextErrors } from '@/components';
+import { Box, Card, Text, TextErrors } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
 import { ModalAddMembers } from '@/features/teams/member-add';
 import { Role, Team } from '@/features/teams/team-management';
 
 import { useTeamAccesses } from '../api';
+import IconMagnifyingGlass from '../assets/icon-magnifying-glass.svg';
 import { PAGE_SIZE } from '../conf';
 import { Access } from '../types';
 
@@ -43,11 +48,28 @@ const defaultOrderingMapping: Record<string, string> = {
  */
 function formatSortModel(
   sortModel: SortModelItem,
-  mapping = defaultOrderingMapping,
-) {
+  mapping: Record<string, string> = defaultOrderingMapping,
+): string {
   const { field, sort } = sortModel;
   const orderingField = mapping[field] || field;
   return sort === 'desc' ? `-${orderingField}` : orderingField;
+}
+
+const MS_DEBOUNCE = 500;
+export function useDebounce(value: string, delay: number = MS_DEBOUNCE) {
+  const [debouncedValue, setDebouncedValue] = useState(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setDebouncedValue(value);
+    }, delay);
+
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return debouncedValue;
 }
 
 export const MemberGrid = ({ team, currentRole }: MemberGridProps) => {
@@ -57,16 +79,33 @@ export const MemberGrid = ({ team, currentRole }: MemberGridProps) => {
   const pagination = usePagination({
     pageSize: PAGE_SIZE,
   });
+
+  const searchMemberValidationSchema = z.object({
+    query: z.string().optional(),
+  });
+
+  const methods = useForm<{ query: string }>({
+    delayError: 0,
+    defaultValues: {
+      query: '',
+    },
+    mode: 'onChange',
+    reValidateMode: 'onChange',
+    resolver: zodResolver(searchMemberValidationSchema),
+  });
+
   const [sortModel, setSortModel] = useState<SortModel>([]);
   const [accesses, setAccesses] = useState<Access[]>([]);
   const { page, pageSize, setPagesCount } = pagination;
 
+  const membersQuery = useDebounce(methods.watch('query'));
   const ordering = sortModel.length ? formatSortModel(sortModel[0]) : undefined;
 
   const { data, isLoading, error } = useTeamAccesses({
     teamId: team.id,
     page,
     ordering,
+    query: membersQuery,
   });
 
   useEffect(() => {
@@ -104,21 +143,51 @@ export const MemberGrid = ({ team, currentRole }: MemberGridProps) => {
 
   return (
     <>
-      {currentRole !== Role.MEMBER && (
-        <Box $margin={{ all: 'big', bottom: 'small' }} $align="flex-end">
-          <Button
-            aria-label={t('Add members to the team')}
-            style={{
-              width: 'fit-content',
-              minWidth: '8rem',
-              justifyContent: 'center',
-            }}
-            onClick={() => setIsModalMemberOpen(true)}
-          >
-            {t('Add a member')}
-          </Button>
+      <Box
+        $display="flex"
+        $direction="column"
+        $margin={{ all: 'big', bottom: 'small' }}
+        $position="relative"
+      >
+        <Text
+          $theme="primary"
+          $weight="600"
+          $size="1.3rem"
+          $margin={{ bottom: 'big' }}
+        >
+          {t('Group members')}
+        </Text>
+        <Box
+          $display="flex"
+          $direction="row"
+          $justify="space-between"
+          $align="center"
+          $gap="1rem"
+          $css={`
+            & > * {
+              flex: 0.23 0 auto;
+            }
+          `}
+        >
+          <Input
+            label={t('Filter member list')}
+            rightIcon={<IconMagnifyingGlass />}
+            {...methods.register('query')}
+          />
+          {currentRole !== Role.MEMBER && (
+            <Button
+              aria-label={t('Add members to the team')}
+              style={{
+                minWidth: '8rem',
+                maxWidth: 'fit-content',
+              }}
+              onClick={() => setIsModalMemberOpen(true)}
+            >
+              {t('Add a member')}
+            </Button>
+          )}
         </Box>
-      )}
+      </Box>
       <Card
         $padding={{ bottom: 'small' }}
         $margin={{ all: 'big', top: 'none' }}